requirements.txt
torch
numpy
pandas
sklearn
scipy 
potpourri
matplotlib